refactor(llama): split forward_layer_batched_decode + post_attn + unified_forward_layer to peer file (Phase 3b)
Address audit's #6 large-file priority: llama_family.rs at 4305 lines.
The 3 batched/unified forward methods accounted for 1652 lines (38%
of the file):

  llama_family.rs:                    4305 → 2652 lines (-1653)
  llama_family_forward_batched.rs:    NEW   1677 lines

Methods moved (as a separate `impl<B: MoeLlmBackend> LlamaFamilyModel<B>`
block in the peer file):
  - forward_layer_batched_decode  (~790 lines, the big batched decode)
  - forward_layer_batched_decode_post_attn  (~570 lines)
  - unified_forward_layer  (~290 lines)

Required visibility bumps to enable cross-file impl block (Rust allows
multiple impl blocks for the same type across crate files):
  - profiling AtomicU64 statics: `static` → `pub(crate) static`
  - SINGLE_ITEM_GRAPH_KEY const → `pub(crate)`
  - 3 fields on LlamaFamilyModel struct (unified_graph_*) → `pub(crate)`
  - ensure_unified_scratch helper method → `pub(crate)`
  - moved methods themselves → `pub(crate)` (so other impl blocks call them)

No behavior change. Both files share the LlamaFamilyModel<B: MoeLlmBackend>
type. Module wiring: `pub mod llama_family_forward_batched;` in models/mod.rs.

Acceptance gates:
  cargo check --workspace --all-targets    clean
  cargo fmt --all -- --check               clean
  RUSTFLAGS=-D warnings cargo check        clean
  Metal smoke (Qwen3-0.6B, 3×128 tok):
    Phase 2.5 baseline: 60.5 tok/s
    Phase 3a:           61.2 tok/s
    Phase 3a+b:         61.2 tok/s (no regression)
  CUDA Vast 4090 (Qwen3-30B-A3B-GPTQ-Int4 c=4 IST-DASLab Marlin path):
    57.6 tok/s aggregate, TPOT 66.65ms — completes without panic
    (this exercises both the moved llama_family forward batched path
    AND the Phase 3a moe/forward.rs free fns, since 30B-A3B layers
    delegate FFN to the moe forward dispatchers)

Combined Phase 3a + 3b file-split scorecard:
  qwen3_moe.rs:    3579 → 2950 (-629; from Phase 3a)
  llama_family.rs: 4305 → 2652 (-1653; this commit)
  Total reduction in the two largest model files: -2282 lines.
  New focused files: moe/forward.rs (657) + llama_family_forward_batched.rs (1677).
